Move Chinese example extractor code to a separate file
